old/requirements.txt
certifi
chardet
idna
oauthlib
PySocks
requests
requests-oauthlib
six
urllib3
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-storage
google-resumable-media
googleapis-common-protos
google-cloud-logging
bs4
lxml